requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.1
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
python-telegram-bot==20.3
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
